app.py
import streamlit as st
import streamlit.components.v1 as components
import os
import tempfile
import zipfile
import requests
import numpy as np
import re
import io
import time
from concurrent.futures import ThreadPoolExecutor, as_completed
from pathlib import Path
from openpyxl import load_workbook
from docx import Document
from docx.table import Table
from docx.text.paragraph import Paragraph
from sklearn.cluster import AgglomerativeClustering
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

# ==========================================
# 0. 구글 애드센스 광고 삽입 함수
# ==========================================
def show_adsense():
    """
    구글 애드센스 코드를 삽입하는 영역입니다.
    애드센스 승인 후 발급받은 HTML/JS 코드를 아래 html_code 문자열 안에 넣으세요.
    """
    html_code = """
    <div style="text-align: center; margin: 20px 0; padding: 10px; background-color: #f0f2f6; border-radius: 5px; color: #888;">
        <p style="font-size: 12px; margin: 0;">Google AdSense Advertisement Area</p>
    </div>
    """
    components.html(html_code, height=100)

# ==========================================
# 1. 환경 설정 및 세션 초기화
# ==========================================
st.set_page_config(page_title="3GPP AI Analyzer Pro", page_icon="📡", layout="wide")

# 세션 상태(Session State) 초기화
if "authenticated" not in st.session_state:
    st.session_state.authenticated = False
if "log_text" not in st.session_state:
    st.session_state.log_text = ""
if "process_done" not in st.session_state:
    st.session_state.process_done = False
if "out1_bio" not in st.session_state:
    st.session_state.out1_bio = None
if "out2_bio" not in st.session_state:
    st.session_state.out2_bio = None
if "extracted_data" not in st.session_state:
    st.session_state.extracted_data = [] # 상세 분석을 위해 텍스트 데이터를 저장하는 리스트

def append_log(text):
    st.session_state.log_text += f"{text}\n"

# (기존 네트워크 및 PIN 함수 생략 없이 유지 - 지면상 간략화, 실제로는 기존 코드의 fetch_remote_pin 등 그대로 유지)
INTERNAL_PROXY = {"http": "http://1