requirements.txt
streamlit
openpyxl
requests
python-docx
scikit-learn
numpy